requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
idna==2.10
nose==1.3.7
numpy==1.19.2
pandas==1.1.2
pycodestyle==2.6.0
pyodbc==4.0.30
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.25.10
